requirements.txt
click==4.0
Flask==0.10.1
flask-hookserver==0.1.3
future==0.14.3
gunicorn==19.3.0
ipaddress==1.0.7
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
redis==2.10.3
requests==2.6.2
rq==0.5.2
Werkzeug==0.10.4
